+++ src/main/java/kr/dogfoot/hwpxlib/object/common/baseobject/WidthAndHeight.java
@@ -3,31 +3,31 @@ package kr.dogfoot.hwpxlib.object.common.baseobject;
 import kr.dogfoot.hwpxlib.object.common.HWPXObject;
 
 public abstract class WidthAndHeight<ChildType> extends HWPXObject {
-    private Integer width;
-    private Integer height;
+    private Long width;
+    private Long height;
 
-    public Integer width() {
+    public Long width() {
         return width;
     }
 
-    public void width(Integer width) {
+    public void width(Long width) {
         this.width = width;
     }
 
-    public ChildType widthAnd(Integer width) {
+    public ChildType widthAnd(Long width) {
         this.width = width;
         return (ChildType) this;
     }
 
-    public Integer height() {
+    public Long height() {
         return height;
     }
 
-    public void height(Integer height) {
+    public void height(Long height) {
         this.height = height;
     }
 
-    public ChildType heightAnd(Integer height) {
+    public ChildType heightAnd(Long height) {
         this.height = height;
         return (ChildType) this;
     }
